pyproject.toml
[project]
name = "modelapi"
version = "0.1.0"
description = "AI Model Serving API - FastAPI-based stateless inference service"
readme = "README.md"
requires-python = ">=3.13"

dependencies = [
    # Web Framework
    "fastapi[standard]",
    "uvicorn[standard]",
    "pydantic-settings",
    "python-dotenv",
    "python-multipart",

    # Machine Learning
    "torch",
    "pytorch-lightning",
    "faiss-cpu",

    # Data Processing
    "pandas",
    "numpy",
    "scipy",

    # Video Processing
    "opencv-python",
    "dlib",

    # Serialization
    "joblib",
]

[project.optional-dependencies]
dev = [
    "ruff",
]

[project.scripts]
modelapi = "app.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.ruff]
line-length = 88
target-version = "py313"
exclude = ["app/models", "app/preprocessing"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.uv]
environments = ["sys_platform == 'linux'", "sys_platform == 'win32'"]
